extract SEARCH_HELPERS
